tests: add and update tests

+++ src/user/services/user.service.ts
@@ -1,48 +1,61 @@
 import { UserRepository } from "../repositories/user.repository";
 import { User } from "../models/user.model";
+import { Types } from "mongoose";
 
 export class UserService {
   constructor(private readonly userRepository: UserRepository) {}
 
-  async getAllUsers(): Promise<User[]> {
+  async getAll(): Promise<User[]> {
     try {
       const users = await this.userRepository.getAll();
 
       return users;
     } catch (error) {
-      throw new Error(error);
+      throw new Error("Error");
     }
   }
 
-  async getUserById(id: string): Promise<User> {
+  async getById(id: string): Promise<User> {
     try {
-      return this.userRepository.getById(id);
+      if (!Types.ObjectId.isValid(id)) {
+        const user = await this.userRepository.getById(id);
+        return user;
+      }
+      throw new Error(`Invalid id: ${id}`);
     } catch (error) {
-      throw new Error(error);
+      throw new Error("Error");
     }
   }
 
-  async createUser(user: User): Promise<User> {
+  async create(user: User): Promise<User> {
     try {
       return this.userRepository.create(user);
     } catch (error) {
-      throw new Error(error);
+      throw new Error("Error");
     }
   }
 
   async update(id: string, user: User): Promise<User> {
     try {
-      return this.userRepository.update(id, user);
+      if (!Types.ObjectId.isValid(id)) {
+        const updatedUser = await this.userRepository.update(id, user);
+        return updatedUser;
+      }
+      throw new Error(`Invalid id: ${id}`);
     } catch (error) {
-      throw new Error(error);
+      throw new Error("Error");
     }
   }
 
   async delete(id: string): Promise<User> {
     try {
-      return this.userRepository.delete(id);
+      if (!Types.ObjectId.isValid(id)) {
+        const deletedUser = await this.userRepository.delete(id);
+        return deletedUser;
+      }
+      throw new Error(`Invalid id: ${id}`);
     } catch (error) {
-      throw new Error(error);
+      throw new Error("Error");
     }
   }
 }
